# Remove redundant coversion of `CharSequence` into `Array` in `Character.codePoint{at, before}` methods